# api/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "version": "2.0.0",
        "title": "Quick Dashboard API",
        "description": "Manages data for the Dashboard",
        "contact": {
            "name": "Benjamin Weare"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://localhost:8000",
            "description": "Local Development"
        },
        {
            "url": "https://intranet-api.local",
            "description": "Hosted Internally"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "APIKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-KEY"
            }
        },
        "schemas": {
...
            "fridayDistance": {
                "type": "string",
                "enum": ["today", "all"]
            },
            "fridayAction": {
                "type": "string",
                "enum": ["add", "remove"]
            },
            "fridayType": {
                "type": "string",
                "enum": ["win", "loss"]
            },
            "error": {
                "required": ["message"],
                "properties": {
                    "message": { "type": "string" }
                }
            },
...
            "sync": { "properties": { "message": { "type": "string" } } },
            "loginRequest": {
                "properties": { "username": { "type": "string" } }
            },
            "userDetails": {
                "properties": {
                    "username": { "type": "string" },
                    "password": { "type": "string" },
                    "firstname": { "type": "string" },
                    "surname": { "type": "string" },
                    "email": { "type": "string" }
                }
            },
...
        }
    },
    "security": [{ "APIKeyAuth": [] }],
    "paths": {
...
        "/friday": {
            "x-swagger-router-controller": "friday",
            "get": {
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/fridayDistance"
                        },
                        "name": "fridayDistance",
                        "in": "query",
                        "required": true
                    }
                ],
                "tags": ["Friday"],
                "summary": "Gets Friday's data",
                "x-eov-operation-handler": "friday",
                "operationId": "getFriday",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sync"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/fridayDistance"
                        },
                        "name": "fridayDistance",
                        "in": "query",
                        "required": true
                    }
                ],
                "tags": ["Friday"],
                "summary": "Delete Friday's data (today, all)",
                "x-eov-operation-handler": "friday",
                "operationId": "deleteFriday",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sync"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/fridayAction"
                        },
                        "name": "fridayAction",
                        "in": "query",
                        "required": true
                    },
                    {
                        "schema": {
                            "$ref": "#/components/schemas/fridayType"
                        },
                        "name": "fridayType",
                        "in": "query",
                        "required": true
                    }
                ],
                "tags": ["Friday"],
                "summary": "Patch Friday's data (win, loss)",
                "x-eov-operation-handler": "friday",
                "operationId": "patchFriday",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sync"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth": {
            "x-swagger-router-controller": "auth",
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/loginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "tags": ["Authentication"],
                "summary": "Attempts to retrieve username",
                "x-eov-operation-handler": "auth",
                "operationId": "loginRequest",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sync"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/create": {
            "x-swagger-router-controller": "auth",
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userDetails"
                            }
                        }
                    },
                    "required": true
                },
                "tags": ["Authentication"],
                "summary": "Gets Signed in user information",
                "x-eov-operation-handler": "auth",
                "operationId": "newUser",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sync"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}